refactor: extract mitm response streaming state
