+++ crates/runner/mitm-addon/src/mitm_addon.py
@@ -21,14 +21,15 @@ from mitmproxy.addonmanager import Loader
 
 # --- Sub-module imports ---
 #
-# Usage/body_utils are imported by module (not selective `from X import ...`)
+# Usage/body_utils/response_streaming are imported by module (not selective `from X import ...`)
 # so that:
-#   1. Cross-module calls read as ``usage.X(...)`` / ``body_utils.X(...)``,
+#   1. Cross-module calls read as ``usage.X(...)`` / ``body_utils.X(...)`` /
+#      ``response_streaming.X(...)``,
 #      making the module boundary visible at call sites.
-#   2. Tests can patch ``usage.name`` / ``body_utils.name`` in one place and
-#      it affects both direct callers in those modules and handler callers
-#      here — no mock-placement pitfalls.
+#   2. Tests can patch names on the owning module object and affect all
+#      callers — no mock-placement pitfalls from copied function bindings.
 import body_utils
+import response_streaming
 import usage
 import vendor_check
 from auth import (
@@ -47,11 +48,7 @@ from url_utils import get_original_url
 # starts shadowing our vendored copy; see vendor_check.py for the playbook.
 vendor_check.verify()
 
-# HTTP status boundaries used in response-phase classification.  Also defined
-# in ``usage.py`` for the same reason; kept local because they're RFC-fixed
-# (never drift) and factoring them out for two callers adds no value.
-_HTTP_STATUS_OK_MIN = 200  # inclusive: start of 2xx success range
-_HTTP_STATUS_REDIRECT_MIN = 300  # start of 3xx — doubles as the 2xx exclusive upper bound
+# HTTP status boundaries used in response-phase classification.
 _HTTP_STATUS_UNAUTHORIZED = 401
 _HTTP_STATUS_ERROR_MIN = 400  # inclusive: start of 4xx/5xx error range
 
@@ -350,166 +347,8 @@ def _release_tracked_usage_flow(flow: http.HTTPFlow) -> None:
 
 
 def responseheaders(flow: http.HTTPFlow) -> None:
-    """
-    Enable response streaming with body buffering.
-
-    Uses a callback to stream response data to the client immediately
-    while accumulating a copy in memory (up to ``STREAM_BUFFER_LIMIT``).
-    Once the limit is exceeded, buffering stops but streaming continues
-    uninterrupted.  The buffered body is available in the ``response()``
-    hook via ``flow.metadata["stream_buffer"]``.
-    """
-    if not flow.response:
-        return
-
-    buf = bytearray()
-    state = {"truncated": False, "total_bytes": 0}
-
-    # Set up usage extraction for billable response classes that need body
-    # inspection. The forensic stream_buffer remains capped; billing parsers
-    # consume chunks separately so a large response cannot grow that buffer.
-    sse_parser = None
-    sse_decompressor = None
-    model_json_parser = None
-    model_json_decompressor = None
-    ndjson_parser = None
-    ndjson_decompressor = None
-    x_json_parser = None
-    x_json_decompressor = None
-    firewall_name = flow.metadata.get("firewall_name", "")
-    is_model_provider = firewall_name.startswith("model-provider:")
-    # Platform-billable firewall flag, sourced from vm_info["billableFirewalls"]
-    # via auth.handle_firewall_request.  Gates report_connector_usage (in response())
-    # and the incremental response parsers used for billing payload extraction.
-    is_billable_flow = flow.metadata.get("firewall_billable", False)
-    is_billable_model_provider = is_model_provider and is_billable_flow
-    # X-specific NDJSON stream classification — tied to the x firewall itself,
-    # not to billing.  Kept separate so a future non-x billable connector
-    # doesn't accidentally inherit X stream parsing.
-    is_x_flow = firewall_name == "x"
-
-    # Classify X NDJSON streams early so we can register an incremental parser
-    # and avoid buffering the (potentially multi-GB) stream body.  Only a
-    # cheap path-only check happens here — full request metadata is parsed
-    # later in response() by report_connector_usage.
-    #
-    # Gated on 2xx status so error responses (4xx/5xx on stream endpoints
-    # return JSON, not NDJSON) skip the billing parser and only use the capped
-    # forensic stream buffer.  report_connector_usage already skips non-2xx
-    # responses so no billing record is affected either way.
-    #
-    # Reads ``original_url`` with no fallback — kept consistent with
-    # :func:`usage._parse_x_request_metadata` so the log entry's
-    # ``is_stream`` field cannot diverge from the parser registration
-    # decision.  For any x firewall flow, ``request()`` has already
-    # populated ``original_url`` before ``responseheaders`` fires.
-    is_x_stream = False
-    if is_x_flow and _HTTP_STATUS_OK_MIN <= flow.response.status_code < _HTTP_STATUS_REDIRECT_MIN:
-        stream_path = urllib.parse.urlparse(flow.metadata.get("original_url", "")).path
-        is_x_stream = usage.x.is_stream_path(stream_path)
-
-    if is_billable_model_provider:
-        content_type = flow.response.headers.get("content-type", "")
-        if "text/event-stream" in content_type:
-            parser_fn, usage_dict = usage.create_sse_usage_extractor()
-            sse_parser = parser_fn
-            flow.metadata["model_provider_usage"] = usage_dict
-            sse_decompressor = body_utils.create_stream_decompressor(flow.response.headers)
-        else:
-            extractor = usage.create_model_json_usage_extractor()
-            model_json_parser = extractor.feed
-            flow.metadata["model_json_usage_finish"] = extractor.finish
-            model_json_decompressor = body_utils.create_stream_decompressor(flow.response.headers)
-    elif is_x_stream:
-        parser_fn, ndjson_state = usage.x.create_ndjson_extractor()
-        ndjson_parser = parser_fn
-        # Deliberately NOT "model_provider_usage" — that key would route through
-        # report_model_provider_usage and trigger the model-provider webhook.
-        # x_ndjson_state is only consumed by report_connector_usage.
-        flow.metadata["x_ndjson_state"] = ndjson_state
-        ndjson_decompressor = body_utils.create_stream_decompressor(flow.response.headers)
-    elif (
-        is_x_flow
-        and is_billable_flow
-        and _HTTP_STATUS_OK_MIN <= flow.response.status_code < _HTTP_STATUS_REDIRECT_MIN
-    ):
-        extractor = usage.x.create_json_response_extractor()
-        x_json_parser = extractor.feed
-        flow.metadata["x_json_response_finish"] = extractor.finish
-        x_json_decompressor = body_utils.create_stream_decompressor(flow.response.headers)
-
-    # Buffer cap policy:
-    # - stream_buffer is only for forensic logging / capture and is always
-    #   capped at STREAM_BUFFER_LIMIT.
-    # - Billing extraction uses the incremental parsers above.
-    buf_limit = body_utils.STREAM_BUFFER_LIMIT
-
-    def stream_and_buffer(chunk: bytes) -> bytes:
-        state["total_bytes"] += len(chunk)
-        if not state["truncated"]:
-            remaining = buf_limit - len(buf)
-            if len(chunk) <= remaining:
-                buf.extend(chunk)
-            else:
-                buf.extend(chunk[:remaining])
-                state["truncated"] = True
-        if sse_parser is not None:
-            plaintext = sse_decompressor(chunk) if sse_decompressor else chunk
-            sse_parser(plaintext)
-        elif model_json_parser is not None:
-            plaintext = model_json_decompressor(chunk) if model_json_decompressor else chunk
-            model_json_parser(plaintext)
-        elif ndjson_parser is not None:
-            plaintext = ndjson_decompressor(chunk) if ndjson_decompressor else chunk
-            ndjson_parser(plaintext)
-        elif x_json_parser is not None:
-            plaintext = x_json_decompressor(chunk) if x_json_decompressor else chunk
-            x_json_parser(plaintext)
-        return chunk
-
-    flow.response.stream = stream_and_buffer
-    flow.metadata["stream_buffer"] = buf
-    flow.metadata["stream_buffer_state"] = state
-    flow.metadata["_vm0_response_stream_callback"] = stream_and_buffer
-
-
-def _finalize_model_json_usage(flow: http.HTTPFlow, proxy_log_path: str) -> None:
-    finish = flow.metadata.pop("model_json_usage_finish", None)
-    if finish is None:
-        return
-    flow.metadata["_model_json_usage_finalized"] = True
-    usage_result, error = finish()
-    if usage_result:
-        flow.metadata["model_provider_usage"] = usage_result
-        return
-    if error:
-        log_proxy_entry(
-            proxy_log_path,
-            "warn",
-            "Model provider JSON usage extraction failed",
-            type="usage_event",
-            error=error,
-        )
-
-
-def _finalize_x_json_state(flow: http.HTTPFlow) -> None:
-    finish = flow.metadata.pop("x_json_response_finish", None)
-    if finish is None:
-        return
-    state, error = finish()
-    if error:
-        state["parse_error"] = error
-    flow.metadata["x_json_state"] = state
-
-
-def _release_response_stream_state(flow: http.HTTPFlow) -> None:
-    stream_callback = flow.metadata.pop("_vm0_response_stream_callback", None)
-    flow.metadata.pop("stream_buffer", None)
-    flow.metadata.pop("stream_buffer_state", None)
-    flow.metadata.pop("model_json_usage_finish", None)
-    flow.metadata.pop("x_json_response_finish", None)
-    if stream_callback is not None and flow.response and flow.response.stream is stream_callback:
-        flow.response.stream = False
+    """Install response stream buffering and incremental body parsers."""
+    response_streaming.configure_response_stream(flow)
 
 
 def _track_usage_flow(fn):
@@ -525,7 +364,7 @@ def _track_usage_flow(fn):
         try:
             return fn(flow, *args, **kwargs)
         finally:
-            _release_response_stream_state(flow)
+            response_streaming.release_response_stream_state(flow)
             _release_tracked_usage_flow(flow)
 
     return wrapper
@@ -607,7 +446,7 @@ def response(flow: http.HTTPFlow) -> None:
 
         log_network_entry(network_log_path, log_entry)
 
-    _finalize_model_json_usage(flow, proxy_log_path)
+    response_streaming.finalize_model_json_usage(flow, proxy_log_path)
 
     # Report proxy-extracted usage for model provider responses.
     # For non-streaming responses, fall back to extracting usage from the
@@ -631,7 +470,7 @@ def response(flow: http.HTTPFlow) -> None:
     usage.report_model_provider_usage(flow, run_id)
 
     # Billable connector usage observation (issue #9504, stage 0).
-    _finalize_x_json_state(flow)
+    response_streaming.finalize_x_json_state(flow)
     usage.report_connector_usage(flow, run_id)
 
     # Invalidate firewall header cache on 401 so next request gets fresh headers.
